Java Programming
Write a method validName(String name) which check if the name is valid name
and return the name, throw InputMismatchException if invalid, In the main() method ask the user to enter a full name and check validity
through validName() method and add try and catch block around appropriate
statements. A valid name contains only alphabets and white space.